requirements.txt
absl-py==2.1.0
amqp==5.2.0
astunparse==1.6.3
billiard==4.2.0
blinker==1.7.0
boto3==1.34.21
botocore==1.34.21
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
etils==1.6.0
Flask==3.0.0
flatbuffers==23.5.26
flax==0.7.5
fsspec==2023.12.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.3
jmespath==1.0.1
keras==2.15.0
kombu==5.3.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
nest-asyncio==1.5.9
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint==0.4.4
packaging==23.2
pandas==2.1.4
prompt-toolkit==3.0.43
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
scipy==1.11.4
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
tensorflowjs==4.16.0
tensorstore==0.1.45
termcolor==2.4.0
toolz==0.12.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0
